Release 2.1.0.14
The following is a new app version 2.1.0.14 is to upgrade existing app framework to include new libraries. The following app have a basic structure to do authentication and App Protection SDK. Readme has been updated to allow the App Protection Policy to be created. Will be releasing this version to Google Manage Play.